backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
supabase>=1.0.0
tiktoken>=0.3.0
numpy>=1.24.0
httpx>=0.24.0
python-multipart>=0.0.6
crawl4ai>=0.1.0  # Advanced web scraping framework
# Azure OpenAI specific packages
azure-identity>=1.12.0  # For Azure authentication (if needed)
azure-core>=1.26.0      # Core functionality for Azure SDK
